requirements.txt
tinydb==4.5.0
python-telegram-bot==13.12
openai==0.10.0